Big Sorting
After doing the brute force method which was wayyyy to slow, solved it by storing each value into a dictionary with the key as the length of the integer and the value being a list for each int that fits that length. Then, we just extend a list by each sorted sub list starting with the smallest key.